# Optimizations for StdWriter