@calavera 2016-03-21
I'm not completely sure we should change this. Yes, the Mutex protects the underlying writer object from this copier, but it doesn't protect it when there are several copiers writing concurrently.

> Writer in two steps, one for the header and one for the actual data.

This is exactly what caused us to release 1.10.3. NOT LGTM.