@cpuguy83 2016-03-21
Ok, instead I've updated this to implement `io.ReaderFrom` so that `io.Copy` can use this 0 alloc function instead of 2 allocs per call to `Write`.
It does not require setting a mutex or storing the header on the writer itself. Instead we just create the header at the beginning of `ReadFrom` before it starts looping.