@MHBauer 2016-03-21
Is there a place to include the benchmark code as well?