@cpuguy83 2016-03-29
This is updated and I tweaked it a bit.
The mutex in the previous iteration was only used for two things:

1. modifying the header (which supposedly reduced allocs)
2. writing to a single shared buffer between writes.

1 was not actually accurate, it was not saving any allocs or improving things at all, so instead it just creates the header on each `write` call

2 Will instead use a sync.Pool to grab a buffer so we don't need a mutex on the writer.

I removed the implementation of `io.ReaderFrom` since it was not any more performant than the new `Write` implementation anyway.

This PR boosts throughput of something like `docker run -v /1G.file:/1G.file busybox cat /1G.file` from 200-250MB/sec to 325-375MB/sec